✏️ extracted showcase body component
