+++ pages/showcase.js
@@ -4,21 +4,29 @@ import Image from '../components/Image';
 import { sortedProjects } from '../companies-manifest';
 import Nav from '../components/Nav';
 import Footer from '../components/Footer';
-import { Github, Bitbucket, Gitlab, Git } from 'styled-icons/fa-brands';
-import { Globe } from 'styled-icons/fa-solid/Globe';
 import { withRouter } from 'next/router';
 import WithIsScrolled from '../components/WithIsScrolled';
 import { generateShowcaseUrl } from '../components/Slider/ShowcaseLink';
 import { CSSTransition, TransitionGroup } from 'react-transition-group';
 import { mobile, phone } from '../utils/media';
 import Navigation from '../components/Slider/Navigation';
+import ShowcaseBody from '../components/Slider/ShowcaseBody';
 
 const Container = styled.div`
   overflow-x: hidden;
 
   * {
-    margin-top: 0;
     font-family: Avenir Next;
+  }
+
+  h1,
+  h2,
+  h3,
+  h4,
+  h5,
+  h6,
+  p {
+    margin-top: 0;
   }
 
   h1 {
@@ -113,76 +121,6 @@ const BodyWrapper = styled.div`
 const Slide = styled(Image)`
   border-radius: 12px;
   box-shadow: 0 32px 48px rgba(0, 0, 0, 0.12);
-`;
-
-const SlideMeta = styled.div`
-  display: flex;
-  justify-content: space-between;
-  margin-top: 48px;
-
-  ${phone(css`
-    position: relative;
-    top: -86px;
-    flex-direction: column-reverse;
-  `)}
-
-  em {
-    color: rgb(219, 112, 147);
-    font-style: normal;
-  }
-`;
-
-const SlideMetaLinks = styled.div`
-  display: flex;
-  flex-direction: column;
-  margin-left: 64px;
-  margin-bottom: 24px;
-  flex-shrink: 0;
-  padding-top: 16px;
-  align-items: flex-end;
-
-  ${phone(css`
-    margin-left: 0;
-    justify-content: center;
-    flex-direction: row;
-
-    span {
-      display: none;
-    }
-  `)}
-`;
-
-const SlideMetaLink = styled.a`
-  display: flex;
-  align-items: center;
-  justify-content: flex-end;
-  margin-bottom: 8px;
-  margin-right: -16px;
-  padding: 4px 16px;
-  border-radius: 8px;
-
-  ${phone(css`
-    display: inline-flex;
-    justify-content: center;
-    margin-right: 0;
-    margin-bottom: 0;
-    padding: 16px;
-    border-radius: 50%;
-  `)}
-
-  &:hover {
-    background-color: rgba(0, 0, 0, 0.05);
-  }
-
-  svg {
-    display: inline-block;
-    height: 24px;
-    margin-left: 8px;
-
-    ${phone(css`
-      margin-left: 0;
-    `)}
-  }
 `;
 
 const getSlide = childIndex => keyframes`
@@ -318,13 +256,6 @@ class ArrowEvents extends React.Component {
   }
 }
 
-const RepoIcon = ({ url }) => {
-  if (url.indexOf('github') > -1) return <Github />;
-  if (url.indexOf('bitbucket') > -1) return <Bitbucket />;
-  if (url.indexOf('gitlab') > -1) return <Gitlab />;
-  return <Git />;
-};
-
 const Showcase = ({ router }) => {
   const { item } = router.query;
   const { currentSlide, previousSlide, nextSlide } = calculateSlides(Object.keys(sortedProjects), item);
@@ -387,27 +318,7 @@ const Showcase = ({ router }) => {
             <Navigation prev={previousSlide} next={nextSlide} />
             <BodyWrapper>
               <InsetWrapper>
-                <SlideMeta>
-                  <div>
-                    <h1>{title}</h1>
-                    <p>
-                      <em>By {owner}</em>
-                    </p>
-                    {description && <p>{description}</p>}
-                  </div>
-                  <SlideMetaLinks>
-                    <SlideMetaLink href={link} target="_blank">
-                      <span>Go to website</span>
-                      <Globe />
-                    </SlideMetaLink>
-                    {repo && (
-                      <SlideMetaLink href={repo} target="_blank">
-                        <span>Go to repository</span>
-                        <RepoIcon url={repo} />
-                      </SlideMetaLink>
-                    )}
-                  </SlideMetaLinks>
-                </SlideMeta>
+                <ShowcaseBody title={title} description={description} owner={owner} link={link} repo={repo} />
               </InsetWrapper>
             </BodyWrapper>
           </Wrapper>
